Refactor into commands execution and dequarantine
